requirements.txt
alembic==1.7.4
attrs==21.2.0
autopep8==1.5.7
black==21.9b0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.1
importlib-resources==5.3.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.5
MarkupSafe==2.0.1
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.9.0
platformdirs==2.4.0
pluggy==1.0.0
psycopg2-binary==2.9.1
py==1.10.0
pycodestyle==2.8.0
PyJWT==2.3.0
pyparsing==2.4.7
pytest==6.2.5
redis==3.5.3
regex==2021.10.8
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.25
SQLAlchemy-Utils==0.37.8
toml==0.10.2
tomli==1.2.1
typed-ast==1.4.3
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0
